Move notifier and resultstracker module in a separate package under the st2action package and update affected tests.
Note: Eventually we should those to a separate standalone package outside
st2actions.
