+++ st2actions/st2actions/cmd/st2notifier.py
@@ -6,8 +6,8 @@ from st2common import log as logging
 from st2common.service_setup import setup as common_setup
 from st2common.service_setup import teardown as common_teardown
 from st2actions import config
-from st2actions import notifier
 from st2actions import scheduler
+from st2actions.notifier import notifier
 
 
 LOG = logging.getLogger(__name__)
