+++ st2actions/st2actions/cmd/st2resultstracker.py
@@ -6,7 +6,7 @@ from st2common import log as logging
 from st2common.service_setup import setup as common_setup
 from st2common.service_setup import teardown as common_teardown
 from st2actions import config
-from st2actions import resultstracker
+from st2actions.resultstracker import resultstracker
 
 
 LOG = logging.getLogger(__name__)
